chalice/requirements.txt
attrs==17.4.0
botocore==1.10.61
certifi==2018.4.16
chalice==1.5.0
chardet==3.0.4
click==6.7
docutils==0.14
enum34==1.1.6
httpie==0.9.9
idna==2.7
Jinja2==2.10
jmespath==0.9.3
Mako==1.0.7
MarkupSafe==1.0
pkg-resources==0.0.0
pycryptodome==3.6.4
Pygments==2.2.0
python-dateutil==2.7.3
requests==2.19.1
six==1.11.0
typing==3.6.4
urllib3==1.23
